requirements.txt
pygame
matplotlib
ipython
python-dotenv